Move get_template_source into input.rs
It's only used in this file.
